[#109] fix: 리마인드가 존재하지않는 카테고리 삭제

+++ linkmind/src/main/java/com/app/toaster/service/category/CategoryService.java
@@ -74,8 +74,9 @@ public class CategoryService {
             categoryRepository.decreasePriorityNextDeleteCategory(categoryId, category.getPriority());
 
             Reminder timer = timerRepository.findByCategory_CategoryId(categoryId);
-
-            timerRepository.delete(timer);
+            if(timer != null)
+                timerRepository.delete(timer);
+            categoryRepository.delete(category);
         }
 
     }
